% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{build_py}
\alias{build_py}
\title{Little helper function to construct the Python-skeleton}
\usage{
build_py(qgis_env = set_env())
}
\arguments{
\item{qgis_env}{Environment settings containing all the paths to run the QGIS
API. For more information, refer to \code{\link{set_env}}.}
}
\description{
This helper function simply constructs the Python-skeleton 
  necessary to run the QGIS-Python API.
}
\examples{
\dontrun{
build_py()
}
}
\author{
Jannes Muenchow
}
\keyword{internal}

