% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{check_for_server}
\alias{check_for_server}
\title{Check if RQGIS is loaded on a server}
\usage{
check_for_server()
}
\description{
Performs cross-platform (Unix, Windows) and OS (Debian/Ubuntu) checks for a server infrastructure
}
\author{
Patrick Schratz
}
\keyword{internal}
