% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{execute_cmds}
\alias{execute_cmds}
\title{Building and executing cmd and Python scripts}
\usage{
execute_cmds(processing_name = "", params = "", qgis_env = set_env(),
  intern = FALSE)
}
\arguments{
\item{processing_name}{Name of the function from the processing library that
should be used.}

\item{params}{Parameter to be used with the processing function.}

\item{qgis_env}{Environment containing all the paths to run the QGIS API. For
more information, refer to \link{\code{set_env}}.}

\item{intern}{Logical which indicates whether to capture the output of the 
command as an \code{R} character vector (see also
\code{\link[base]{system}}.}
}
\description{
This helper function constructs the batch and Python scripts
  which are necessary to run QGIS from the command line.
}
\author{
Jannes Muenchow
}

