% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{find_root}
\alias{find_root}
\title{Looking for OSGEO on your system}
\usage{
find_root(root_name = "OSGeo4W")
}
\arguments{
\item{root_name}{Name of the folder where QGIS, SAGA, GRASS, etc. is
installed.}
}
\description{
\code{find_root} looks for OSGeo on your system under C:, C:/Program
  Files and C:/Program Files (x86). So far, this function is only available
  for Windows.
}
\examples{
find_root()
}
\author{
Jannes Muenchow
}

