% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{get_html_help}
\alias{get_html_help}
\title{Access the QGIS/GRASS online help for a specific function}
\usage{
get_html_help(alg, qgis_env = set_env())
}
\arguments{
\item{alg}{The name of the algorithm for which you wish to retrieve arguments
and default values.}

\item{qgis_env}{Environment containing all the paths to run the QGIS API. For
more information, refer to \code{\link{set_env}}.}
}
\value{
The function opens your default web browser and displays the help for
  the specified algorithm.
}
\description{
\code{get_html_help} opens the online help for a specific function.
  This is the help you also encounter in the QGIS GUI. Please note that you
  are referred to the GRASS documentation in the case of GRASS algorithms.
}
\note{
\code{get_html_help} only works with a working Internet connection.
}
\examples{
# QGIS example
get_html_help(alg = "qgis:addfieldtoattributestable")
# GRASS example
get_html_help(alg = "grass:v.overlay")
}
\author{
Jannes Muenchow, Victor Olaya, QGIS core team
}

