% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{get_options}
\alias{get_options}
\title{Get options of parameters for a specific GIS option}
\usage{
get_options(algorithm_name = "", osgeo4w_root = ifelse(Sys.info()["sysname"]
  == "Windows", find_root(), NULL))
}
\arguments{
\item{algorithm_name}{Name of the GIS function for which options should be
returned.}

\item{osgeo4w_root}{Path to OSGeo4W on your system.}
}
\description{
\code{get_options} lists all available parameter options for
  the required GIS function.
}
\examples{
get_options(algorithm_name = "saga:slopeaspectcurvature")
}
\author{
Jannes Muenchow, QGIS devleoper team
}

