% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{get_output_names}
\alias{get_output_names}
\title{Get output function argument for a specific  QGIS geoalgorithm}
\usage{
get_output_names(alg, qgis_env = set_env())
}
\arguments{
\item{alg}{The name of the algorithm for which one wishes to retrieve
the output arguments.}

\item{qgis_env}{Environment settings containing all the paths to run the QGIS
API. For more information, refer to \code{\link[=set_env]{set_env()}}.}
}
\description{
The function retrieves the function arguments corresponding to
the output names for a specific QGIS geoalgorithm.
}
\examples{
\dontrun{
get_output_names("grass7:r.slope.aspect")
}
}
\author{
Jannes Muenchow
}
\keyword{internal}
