% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{open_app}
\alias{open_app}
\title{Open a QGIS application}
\usage{
open_app(qgis_env = set_env())
}
\arguments{
\item{qgis_env}{Environment settings containing all the paths to run the QGIS
API. For more information, refer to \code{\link[=set_env]{set_env()}}.}
}
\value{
The function enables a 'tunnel' to the Python QGIS API.
}
\description{
\code{open_app} first sets all the correct paths to the QGIS Python
binary, and secondly opens a QGIS application while importing the most
common Python modules.
}
\examples{
\dontrun{
open_app()
}
}
\author{
Jannes Muenchow
}
