% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{pass_args}
\alias{pass_args}
\title{Helper function to specify QGIS geoalgorithm parameters the R way}
\usage{
pass_args(alg, ..., params = NULL, qgis_env = set_env())
}
\arguments{
\item{alg}{The name of the geoalgorithm to use.}

\item{...}{Triple dots can be used to specify QGIS geoalgorithm arguments as
R named arguments.}

\item{params}{Parameter argument list for a specific geoalgorithm, see
\code{\link[=get_args_man]{get_args_man()}} for more details. Please note that you can either specify
R arguments directly via the triple dots (see above) or via the parameter
argument list. However, you may not mix the two methods.}

\item{qgis_env}{Environment containing all the paths to run the QGIS API. For
more information, refer to \code{\link[=set_env]{set_env()}}.}
}
\value{
The function returns the complete parameter argument list for a given
QGIS geoalgorithm. The parameters are ordered as expected by the QGIS API.
The list is constructed with the help of \code{\link[=get_args_man]{get_args_man()}} while considering
the R named arguments or the \code{params}-parameter specified by the user as
additional input.
}
\description{
The function lets the user specify QGIS geoalgorithm parameters
as R named arguments. When omitting required parameters, defaults will be
used if available as derived from \code{\link[=get_args_man]{get_args_man()}}.
}
\note{
The function was inspired by \code{\link[rgrass7:doGRASS]{rgrass7::doGRASS()}}.
}
\examples{
\dontrun{
data(dem, package = "RQGIS")
get_usage("grass7:r.slope.aspect")
# 1. using R named arguments
pass_args("grass7:r.slope.aspect", elevation = dem, 
          slope = file.path(tempdir(), "slope.asc"))
# 2. doing the same with a parameter argument list
pass_args("grass7:r.slope.aspect", 
          params = list(elevation = dem, 
                        slope = file.path(tempdir(), "slope.asc")))
}
}
\author{
Jannes Muenchow
}
