% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_draw_bbox.R
\name{qgis_draw_bbox}
\alias{qgis_draw_bbox}
\title{Draw a Rectangle on the QGIS Map Canvas}
\usage{
qgis_draw_bbox()
}
\value{
A \code{bbox} object from the \code{sf} package representing the drawn extent.
}
\description{
Interactively prompts the user to draw a rectangle on the QGIS map canvas
and returns its bounding box.
}
\examples{
\dontrun{
# Draw a bounding box and save the extent to a variable
my_bbox <- qgis_draw_bbox()
print(my_bbox)
}

}
