% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{qgis_load_geom}
\alias{qgis_load_geom}
\title{Load and register spatial objects in QGIS}
\usage{
qgis_load_geom(params, type_name)
}
\arguments{
\item{params}{A parameter-argument list as returned by \code{\link[=pass_args]{pass_args()}}.}

\item{type_name}{A character string containing the QGIS parameter type for
each parameter (boolean, multipleinput, extent, number, vector, raster,
etc.) of \code{params}. The Python method \code{RQGIS.get_args_man} returns a Python
dictionary with one of its elements corresponding to the type_name (see
also the example section).}
}
\value{
The function loads and registers geometry data in a QGIS session
using raster and vector input paths as indicated in the input
parameter-argument list. The function returns the same list, however, it
replaces the original paths to the input raster and vector files on disk by
the layer names in the QGIS session.
}
\description{
The function loads and registers vector and raster files in an
open Python QGIS API session.
}
\examples{
\dontrun{
alg <- "saga:sagawetnessindex"
params <- get_args_man(alg)
params$DEM <- dem
params <- pass_args(alg, params = params)
out <- py_run_string(sprintf("out = RQGIS.get_args_man('\%s')", alg))$out
# make sure to only use input parameters, and not output parameters
params[!out$output] <- qgis_load_geom(params = params[!out$output], 
                                      type_name = out$type_name[!out$output])
}
}
\keyword{internal}
