% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{read_cmds}
\alias{read_cmds}
\title{Read command skeletons}
\usage{
read_cmds(qgis_env = set_env())
}
\arguments{
\item{qgis_env}{Environment containing all the paths to run the QGIS API. For
more information, refer to \link{\code{set_env}}.}
}
\description{
This function simply reads prefabricated Python and batch
  commands.
}
\author{
Jannes Muenchow
}

