% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{reset_path}
\alias{reset_path}
\title{Reset PATH}
\usage{
reset_path(settings)
}
\arguments{
\item{settings}{A list as derived from \code{as.list(Sys.getenv())}.}
}
\description{
Since \code{\link[=run_ini]{run_ini()}} starts with a clean PATH, this function makes
sure to add the original paths to PATH. Note that this function is a
Windows-only function.
}
\author{
Jannes Muenchow
}
