% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{run_ini}
\alias{run_ini}
\title{Reproduce o4w_env.bat script in R}
\usage{
run_ini(qgis_env = set_env())
}
\arguments{
\item{qgis_env}{Environment settings containing all the paths to run the QGIS
API. For more information, refer to \code{\link[=set_env]{set_env()}}.}
}
\value{
The function changes the system settings using \code{\link[base:Sys.setenv]{base::Sys.setenv()}}.
}
\description{
Windows helper function to start QGIS application. Basically,
the code found in all .bat files found in etc/ini (most likely
"C:/OSGEO4~1/etc/ini") is reproduced within R.
}
\examples{
\dontrun{
run_ini()
}
}
\author{
Jannes Muenchow
}
\keyword{internal}
