% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taudem.R
\name{run_taudem}
\alias{run_taudem}
\title{Helper function to run TauDEM}
\usage{
run_taudem(alg, params, qgis_env)
}
\arguments{
\item{alg}{Name of the GIS function to be used (see
\code{\link[=find_algorithms]{find_algorithms()}}).}

\item{params}{A list of geoalgorithm function arguments that should be
used in conjunction with the selected (Q)GIS function (see
\code{\link[=get_args_man]{get_args_man()}}). Please make sure to provide all function
arguments in the correct order. To make sure this is the case, it is
recommended to use the convenience function \code{\link[=get_args_man]{get_args_man()}}.}

\item{qgis_env}{Environment containing all the paths to run the QGIS API.
For more information, refer to \code{\link[=set_env]{set_env()}}.}
}
\description{
\code{run_taudem} calls QGIS-TauDEM algorithms from within R
while passing the corresponding function arguments.
}
\note{
\code{run_taudem} is an internal function which is called by
\code{run_qgis}.
}
\examples{
 \dontrun{
 data("dem")
 alg <- "taudem:d8flowdirections"
 params <- get_args_man(alg = alg, options = TRUE)
 params$`-fel` <- dem
 params$`-p` <- "out_1.tif"
 params$`-sd8` <- "out_2.tif"
 run_qgis(alg = alg, params = params)
 }
}
\keyword{internal}
