% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{saga_support}
\alias{saga_support}
\title{Find the SAGA versions QGIS supports}
\usage{
saga_support(qgis_env = set_env())
}
\arguments{
\item{qgis_env}{Environment settings containing all the paths to run the QGIS
API. For more information, refer to \code{\link{set_env}}.}
}
\value{
The function returns a character vector representing the SAGA version 
  numbers QGIS currently supports.
}
\description{
\code{saga_support} calls via the QGIS Python API the
  \code{SagaAlgorithmProvider} to figure out with which SAGA versions QGIS
  currently is compatible.
}
\examples{
\dontrun{
saga_support()
}
}
\author{
Jannes Muenchow, Victor Olaya, QGIS core team
}

