% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{set_py_config}
\alias{set_py_config}
\title{Set py_config appropiately}
\usage{
set_py_config(dev = NULL, python_version = NULL)
}
\arguments{
\item{dev}{If set to \code{TRUE}, \code{set_env} will use the development version of
QGIS (if available).}

\item{python_version}{Optional use a custom Python version.
Please provide the full path to the respective executable.
This is particulary useful if you want use RQGIS with QGIS 2.99 as we
currently require Python2 for 'LTR' and 'DEV' setting in \code{set_env()}.}
}
\value{
Sets \code{py_config} accordingly.
}
\description{
Helper function to set py_config appropriately or manually
}
\seealso{
\code{\link{set_env}}
}
\author{
Patrick Schratz
}
\keyword{internal}
