% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{setup_linux}
\alias{setup_linux}
\title{Set all Linux paths necessary to start QGIS}
\usage{
setup_linux(qgis_env = set_env())
}
\arguments{
\item{qgis_env}{Environment settings containing all the paths to run the QGIS
API. For more information, refer to \code{\link[=set_env]{set_env()}}.}
}
\value{
The function changes the system settings using \code{\link[base:Sys.setenv]{base::Sys.setenv()}}.
}
\description{
Helper function to start QGIS application under Linux.
}
\examples{
\dontrun{
setup_linux()
}
}
\author{
Jannes Muenchow
}
\keyword{internal}
