% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{setup_win}
\alias{setup_win}
\title{Reproduce o4w_env.bat script in R}
\usage{
setup_win()
}
\arguments{
\item{qgis_env}{Environment settings containing all the paths to run the QGIS
API. For more information, refer to \code{\link[=set_env]{set_env()}}.}
}
\value{
The function changes the system settings using \code{\link[base:Sys.setenv]{base::Sys.setenv()}}.
}
\description{
Windows helper function to start QGIS application by setting all
necessary path especially through running \code{\link[=run_ini]{run_ini()}}.
}
\examples{
\dontrun{
run_ini()
}
}
\author{
Jannes Muenchow
}
\keyword{internal}
