% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{check_apps}
\alias{check_apps}
\title{Checking paths to QGIS applications on Windows}
\usage{
check_apps(osgeo4w_root)
}
\arguments{
\item{osgeo4w_root}{Path to the root directory of the OSGeo4W-installation,
usually C:/OSGeo4W64 or C:/OSGeo4w32.}
}
\value{
The function returns a list with the paths to all the necessary
  QGIS-applications.
 @examples 
\dontrun{
check_apps("C:/OSGeo4W64)
}
}
\details{
\code{check_apps} checks if all the necessary applications (QGIS,
  Python27, Qt4, GDAL, GRASS, msys, SAGA) are installed in the correct
  locations.
}
\author{
Jannes Muenchow, Patrick Schratz
}

