% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{check_apps}
\alias{check_apps}
\title{Checking paths to QGIS applications}
\usage{
check_apps(root, ...)
}
\arguments{
\item{root}{Path to the root directory. Usually, this is 'C:/OSGEO4~1',
'/usr' and '/Applications/QGIS.app/' for the different platforms.}

\item{...}{Optional arguments used in \code{check_apps}. Under Windows,
\code{set_env} passes function argument \code{dev} to \code{check_apps}.}
}
\value{
The function returns a list with the paths to all the necessary
QGIS-applications.
}
\description{
\code{check_apps} checks if software applications necessary to
run QGIS (QGIS and Python plugins) are installed in the correct
locations.
}
\examples{
\dontrun{
check_apps()
}
}
\author{
Jannes Muenchow, Patrick Schratz
}
\keyword{internal}
