% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{find_algorithms}
\alias{find_algorithms}
\title{Find and list available QGIS algorithms}
\usage{
find_algorithms(search_term = "",
  osgeo4w_root = ifelse(Sys.info()["sysname"] == "Windows", find_root(),
  NULL))
}
\arguments{
\item{search_term}{A character to query QGIS functions, i.e. to list only
functions which contain the indicated string.}

\item{osgeo4w_root}{Path to the OSGEO or QGIS folder on your system.}
}
\value{
Python console output will be captured as an R character vector.
}
\description{
\code{find_algorithms} lists or queries all algorithms which
  can be used via the command line and the QGIS API.
}
\details{
Function \code{find_algorithms} simply calls
  \code{processing.alglist} using Python.
}
\examples{
# list all available QGIS algorithms
find_algorithms()
# find a function which adds coordinates
find_algorithms(search_term = "add")
}
\author{
Jannes Muenchow, QGIS developer team
}

