% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{find_algorithms}
\alias{find_algorithms}
\title{Find and list available QGIS algorithms}
\usage{
find_algorithms(search_term = "", name_only = FALSE)
}
\arguments{
\item{search_term}{A character to query QGIS functions, i.e. to list only
functions which contain the indicated string. If empty (\code{""}), the
default, all available functions will be returned.}

\item{name_only}{If \code{TRUE}, the function returns only the name(s) of the
found algorithms. Otherwise, a short function description will be returned
as well (default).}
}
\value{
The function returns QGIS function names and short descriptions as an
R character vector.
}
\description{
\code{find_algorithms} lists or queries all QGIS algorithms
which can be used accessed through the command line.
}
\details{
Function \code{find_algorithms} simply calls
\code{processing.alglist} using Python.
}
\examples{
\dontrun{
# list all available QGIS algorithms on your system
algs <- find_algorithms()
algs[1:15]
# just find all native, i.e. QGIS-algorithms
grep("qgis:", algs, value = TRUE)
# find a function which adds coordinates
find_algorithms(search_term = "add")
}
}
\author{
Jannes Muenchow, Victor Olaya, QGIS core team
}
