% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{get_options}
\alias{get_options}
\title{Get options of parameters for a specific GIS option}
\usage{
get_options(algorithm_name = "", qgis_env = set_env())
}
\arguments{
\item{algorithm_name}{Name of the GIS function for which options should be
returned.}

\item{qgis_env}{Environment containing all the paths to run the QGIS API. For
more information, refer to \link{\code{set_env}}.}
}
\description{
\code{get_options} lists all available parameter options for
  the required GIS function.
}
\details{
Function \code{get_options} simply calls
  \code{processing.algoptions} using Python.
}
\examples{
get_options(algorithm_name = "saga:slopeaspectcurvature")
}
\author{
Jannes Muenchow, QGIS devleoper team
}

