% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{open_grass_help}
\alias{open_grass_help}
\title{Open the GRASS online help}
\usage{
open_grass_help(alg)
}
\arguments{
\item{alg}{The name of the algorithm for which one wishes to retrieve
arguments and default values.}
}
\description{
\code{open_grass_help} opens the GRASS online help for a 
  specific GRASS geoalgorithm.
}
\examples{
\dontrun{
open_grass_help("grass7:r.sunmask")
}
}
\author{
Jannes Muenchow
}
\keyword{export}
\keyword{internal}

