% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{read_cmds}
\alias{read_cmds}
\title{Read command skeletons}
\usage{
read_cmds(osgeo4w_root = find_root())
}
\arguments{
\item{osgeo4w_root}{Path to the OSGeo folder or QGIS folder}
}
\description{
This function simply reads prefabricated Python and batch
  commands.
}
\author{
Jannes Muenchow
}

