% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{run_qgis}
\alias{run_qgis}
\title{Interface to QGIS commands}
\usage{
run_qgis(algorithm = NULL, params = list(), osgeo4w_root = find_root())
}
\arguments{
\item{algorithm}{Name of the GIS function to be used (see
\code{\link{find_algorithms}}).}

\item{params}{A list of function arguments that should be used in conjunction
with the selected GIS function (see \code{\link{get_usage}} and
\code{\link{get_options}}).}

\item{osgeo4w_root}{Path to OSGeo4W on your system.}
}
\description{
\code{run_qgis} is the workhorse of the R-QGIS interface: It
  calls the QGIS API from within R to run QGIS algorithms while passing the
  corresponding function arguments.
}
\details{
This workhorse function calls QGIS via Python (QGIS API) using the
  command line.
}
\examples{
\dontrun{
# find out how a function is called
find_algorithms(search_term = "add")
# find out how it works
get_usage(algorithm_name = "saga:addcoordinatestopoints")
# specify the parameters in the exact same order as listed by get_usage
params <- list(INPUT = "C:/Users/pi37pat/Desktop/test/random_squares.shp",
     OUTPUT = "C:/Users/pi37pat/Desktop/test/qgis_testi2.shp")
run_qgis(algorithm = "saga:addcoordinatestopoints",
         params = params)
}
}
\author{
Jannes Muenchow, QGIS developer team
}

