% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{run_qgis}
\alias{run_qgis}
\title{Interface to QGIS commands}
\usage{
run_qgis(algorithm = NULL, params = list(), qgis_env = set_env())
}
\arguments{
\item{algorithm}{Name of the GIS function to be used (see
\code{\link{find_algorithms}}).}

\item{params}{A list of function arguments that should be used in conjunction
with the selected GIS function (see \code{\link{get_usage}} and
\code{\link{get_options}}).}

\item{qgis_env}{Environment containing all the paths to run the QGIS API. For
more information, refer to \link{\code{set_env}}.}
}
\description{
\code{run_qgis} is the workhorse of the R-QGIS interface: It
  calls the QGIS API from within R to run QGIS algorithms while passing the
  corresponding function arguments.
}
\details{
This workhorse function calls QGIS via Python (QGIS API) using the
  command line. Specifically, it calls \code{processing.runalg}.
}
\examples{
\dontrun{
# set the environment
my_env <- set_env()
# find out how a function is called
find_algorithms(search_term = "add", qgis_env = my_env)
# find out how it works
get_usage(algorithm_name = "saga:addcoordinatestopoints", qgis_env = my_env)
# specify the parameters in the exact same order as listed by get_usage
params <- list(INPUT = "random_squares.shp",
               OUTPUT = "output.shp")
run_qgis(algorithm = "saga:addcoordinatestopoints",
         params = params,
         qgis_env = my_env)
}
}
\author{
Jannes Muenchow, QGIS developer team
}

