% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{set_env}
\alias{set_env}
\title{Retrieve the environment settings to run QGIS from within R}
\usage{
set_env(path = NULL, qgis = NULL, python27 = NULL, qt4 = NULL,
  gdal = NULL, msys = NULL, grass = NULL, saga = NULL)
}
\arguments{
\item{path}{Path to the OSGeo4W-installation (only for machines running on
Windows).}
}
\description{
\code{set_env} tries to find all the paths necessary to run QGIS
  from within R.
}
\details{
If you do not specify function parameter \code{path}, the function 
  looks for \code{qgis.bat}-file on your C: drive. However, this only works 
  if you have used the OSGeo4W-installation. That means, if you installed
  QGIS on your system without using the OSGeo4W-routine, the function might
  still be able to find the QGIS-installation. However, RQGIS will throw an
  error message since \code{check_apps} will not find the dependencies
  necessary to use the Python QGIS API.
}
\examples{
set_env()
}
\author{
Jannes Muenchow
}

